components moved into molecules or organisms
